new/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arabic-reshaper==3.0.0
click==8.3.1
fastapi==0.128.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
pyngrok==7.5.0
python-bidi==0.6.7
python-multipart==0.0.21
PyYAML==6.0.3
qrcode==8.2
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
